Honor MTU size for Ipv4
With IP DF bit set ping was not honoring the MTU size.
Using ping with e.g. "-l 1480" was sending packets with an MTU
of 1508 with the default MTU set to 1500.

Fragmentation is now always performed, based on the vpcs MTU size.
The DF bit is now just prohibiting routers to further fragment them.

Signed-off-by: Alexander Wetzel <alexander@wetzel-home.de>